:recycle: #3 pages/signup.vueのロジックをcomponents配下のコンポーネントに移動
